requirements.txt
transformers==4.48.0
datasets==3.2.0
peft==0.14.0
accelerate==1.2.1
bitsandbytes==0.45.0
PyYAML>=6.0
rouge-score==0.1.2
wandb==0.19.3
dill<0.3.9,>=0.3.0
multiprocess<0.70.17
fsspec<=2024.9.0,>=2023.1.0
requests==2.32.3
python-dotenv==1.0.1
